updater/requirements.txt
async-timeout==4.0.3
beautifulsoup4==4.12.3
blinker==1.4
certifi==2024.7.4
charset-normalizer==3.3.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.14.0
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
pandas==2.2.2
pipenv==2023.12.1
platformdirs==4.2.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
redis==5.0.7
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
spotipy==2.24.0
systemd-python==234
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
virtualenv==20.26.1
virtualenv-clone==0.3.0
wadllib==1.3.6
zipp==1.0.0
